Fixed issues by multiple threads in single field
